Add simpler entity repository with conventions
The new `IEntityRepository` implements some common heuristics
when persisting entities using a fixed `PartitionKey` per entity
type. When entities have a unique identifier as a single property
(i.e. a `Guid`), it's simpler to just use a fixed partition key,
since the entity ID is enough to uniquely access it.

Whether to store all entities in a single table named `Entity` or
not is something that can be decided per entity type (via the
`[Table]` attribute) or per repository instance by specifying it
via the constructor.