Significant refactoring for simplified API and less dependency on reflection
For v2, we want to make it possible to have entities that have no attributes
whatsoever, and instead lambdas can be provided to retrieve rowkey and
partition key values for the target entities.

The existing IEntityPartition was renamed to ITablePartition, since "entity"
is an overloaded term without much meaning. The new name reflects the
fact that it operates on a single partition, as determined by the partition
key used when the object is created.

In addition, instead of directly constructing the implementation objects,
we provide factory classes for both instead, so the code can always
depend on the interfaces instead.