Restrict visibility of implementations constructors
The intended usage pattern for the library is to create the implementations using the `TableRepository.Create` or `TablePartition.Create` factory methods.

The implementations are public so that if a consumer wants to inherit them and provide different defaults for the constructors, they can do that, but users should not be instantiating the implementations directly.

For DI scenarios, the factory method can be used to provide instances, or a such a derived generic class providing default values for the lambdas that are hard to inject (the `Func<T, string>` for partitionKey and rowKey) can be used instead.

Fixes #14