Add an AttributedTableRepository<T> for easy consumption in DI scenarios
In DI scenarios, it's virtually impossible to inject a string with a target table name, or a Func<T, string> for the functions to retrieve a partition key or row key for an entity.

This new implementation uses the same defaults from the factory methods to pass to the base implementation the required arguments, making the DI scenario possible by just requiring a CloudStorageAccount in the container.

Fixes #16