Allow persisting entities as documents
Introduce a document-based repository that allows persisting the entire entity payload to a single column.

Regardless of the serialization strategy used, we persist the type full name and assembly version of the persisted entity, which can be invaluable in data migration scenarios.

In addition to the built-in JSON text serialization, we provide three binary serializers too as separate packages: Bson, MessagePack and Protobuf. The last two require annotating the entity as required by the underlying libraries.

Fixes #24.