Don't duplicate PartitionKey/RowKey properties in storage
Switch to expressions for table repository, so we can inspect the lambda looking for the actual property name being used for partition/row keys. This way we can avoid persisting those as (duplicate) individual columns (since they will already be persisted as partition/row keys).

For document-based storage, we don't need to distinguish, since we persist the entire serialized object, so it's not useful in that case to try to avoid persisting them separately.

Fixes #26