Add support for enhanced filtering for document-based repo
We now need a new document-based specific interface since it will have a new method (EnumerateAsync) which is exclusive for documents. So we rename the old interfaces to `ITableStorage` and `ITableStoragePartition`, which now become the base interfaces for both, and introduce the new `IDocumentRepository` and `IDocumentPartition` which basically add the new enumeration with a filtering expression.

The implementation for documents relies on the underlying `TableQuery`, since we are persisting actual `TableEntity` objects and therefore don't need any of the shenanigans from our POCO repos.

Fixes #37