Allow creating partitions from the same table connection
By exposing the explicit concept of an affinitized table-based storage account+table name, we make it possible to avoid checking for table existence on every new instance of a table repository or partition.

We now also don't fire up a task to check for the table existence, which also minimizes chances of an unhandled background exception.

Closes #155