Properly persist computed colums for same type
We had a bug where we cached the list of properties to persist by type, without regard for the variability in PK/RK names which can be different when persisting the same entity using various key strategies for indexing purposes.

By considering these two properties as part of the cache key, we allow proper persistence as originally intended to avoid duplicating PK/RK props.

Fixes #252