Introduce blob-based storage using ITableStorage<T> API
We modify the binary-based serializers to work on top of Stream instead of byte arrays.

And the new blob-based implementations use the table name as the container name and the partition key as the folder path for the blobs. This allows using document-storage like API without hitting the limitation of column sizes in table storage.